requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Async & Concurrency
aiohttp==3.9.1
httpx==0.25.2
redis[async]==5.0.1
aioredis==2.0.1  # Legacy support

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.4

# Blockchain
web3==6.11.3
eth-account==0.9.0
eth-keys==0.4.0
eth-typing==3.5.2

# Alith SDK (Decentralized AI Agent Framework)
# Install with: pip install alith -U
alith>=0.1.0

# EigenDA Python SDK (Optional - community SDK)
# Install with: pip install git+https://github.com/powerloom/eigenda-py.git
# Note: Official EigenDA uses REST API (implemented in eigenda_client.py)

# LLM Providers
google-generativeai==0.3.1
openai==1.3.5
langchain==0.0.350
langchain-community==0.0.10

requests==2.31.0

# Security Tools
slither-analyzer==0.10.1
PyJWT==2.8.0
cryptography==41.0.7

# Solidity Compilation
py-solc-x==2.0.4
solc-select>=1.1.0  # Solidity compiler version manager

# CLI
click==8.1.7
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Math & Vector Operations
numpy==1.24.3

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

